Fix/split models and schemas (#494)
* fix(B-464): split admin stats model and schema

* fix(B-464): fix missing import of admin-stats schema

* fix(B-464): put AdminDashboardStatsRead in the right _init_

* fix(B-464): split app config model and schema

* fix(B-464): split conversation category model and schema and delete ConversationCategoryCreate

* fix(B-464): split conversation scenario model and schema

* fix(B-464): remove personalization option model

* fix(B-464): delete unused rating model

* fix(B-464): split review model and schema

* fix(B-464): rename scenario_preparation_schema to scenario_preparation

* fix(B-464): clean up scenario_preparation model and schema

* fix(B-464): cleanup session feedback schema and model

* fix(B-464): rename session_feedback schema file

* fix(B-464): split session turn schema and model

* fix(B-464): split session model and schema

* fix(B-464): move paginated session schema to schemas folder

* fix(B-464): split confidence score schema and model

* fix(B-464): delete unused user goal schemas

* fix(B-464): split user_profile model and schema

* fix(B-464): migration script

---------
