+++ backend/alembic/versions/2025_06_20_0012-0a10f61dc9c7_split_db_model_and_pydantic_schemas_and_.py
@@ -1,0 +1,51 @@
+"""split db model and pydantic schemas and deleted rating
+
+Revision ID: 0a10f61dc9c7
+Revises: ccbf2a5fb6a6
+Create Date: 2025-06-20 00:12:33.030494
+
+"""
+
+from collections.abc import Sequence
+from typing import Union
+
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = '0a10f61dc9c7'
+down_revision: Union[str, None] = 'ccbf2a5fb6a6'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('rating')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        'rating',
+        sa.Column('id', sa.UUID(), autoincrement=False, nullable=False),
+        sa.Column('session_id', sa.UUID(), autoincrement=False, nullable=False),
+        sa.Column('user_id', sa.UUID(), autoincrement=False, nullable=False),
+        sa.Column('score', sa.INTEGER(), autoincrement=False, nullable=False),
+        sa.Column('comment', sa.VARCHAR(), autoincrement=False, nullable=False),
+        sa.Column('created_at', postgresql.TIMESTAMP(), autoincrement=False, nullable=False),
+        sa.Column('updated_at', postgresql.TIMESTAMP(), autoincrement=False, nullable=False),
+        sa.ForeignKeyConstraint(
+            ['session_id'], ['session.id'], name=op.f('rating_session_id_fkey'), ondelete='CASCADE'
+        ),
+        sa.ForeignKeyConstraint(
+            ['user_id'], ['userprofile.id'], name=op.f('rating_user_id_fkey'), ondelete='CASCADE'
+        ),
+        sa.PrimaryKeyConstraint('id', name=op.f('rating_pkey')),
+    )
+    # ### end Alembic commands ###
