+++ backend/app/data/__init__.py
@@ -5,7 +5,6 @@ from .dummy_data import (
     get_dummy_app_configs,
     get_dummy_conversation_categories,
     get_dummy_conversation_scenarios,
-    get_dummy_ratings,
     get_dummy_reviews,
     get_dummy_scenario_preparations,
     get_dummy_session_feedback,
@@ -26,7 +25,6 @@ __all__ = [
     'get_dummy_sessions',
     'get_dummy_session_feedback',
     'get_dummy_scenario_preparations',
-    'get_dummy_ratings',
     'get_dummy_user_confidence_scores',
     'get_dummy_app_configs',
     'get_dummy_admin_stats',
