+++ backend/app/data/populate_dummy_data.py
@@ -6,7 +6,6 @@ from app.data import (
     get_dummy_app_configs,
     get_dummy_conversation_categories,
     get_dummy_conversation_scenarios,
-    get_dummy_ratings,
     get_dummy_reviews,
     get_dummy_scenario_preparations,
     get_dummy_session_feedback,
@@ -77,12 +76,6 @@ def populate_data() -> None:
         # Populate Scenario Preparations
         scenario_preparations = get_dummy_scenario_preparations(conversation_scenarios)
         db_session.add_all(scenario_preparations)
-
-        # Populate Ratings
-        ratings = get_dummy_ratings(
-            sessions, conversation_scenarios
-        )  # Pass both sessions and scenarios
-        db_session.add_all(ratings)
 
         # Populate Admin Dashboard Stats
         admin_stats = get_dummy_admin_stats()
