+++ backend/app/models/__init__.py
@@ -1,105 +1,40 @@
-from app.models.admin_dashboard_stats import AdminDashboardStats, AdminDashboardStatsRead
-from app.models.app_config import AppConfig, AppConfigCreate, AppConfigRead, ConfigType
-from app.models.conversation_category import (
-    ConversationCategory,
-    ConversationCategoryCreate,
-    ConversationCategoryRead,
-)
+from app.models.admin_dashboard_stats import AdminDashboardStats
+from app.models.app_config import AppConfig, ConfigType
+from app.models.conversation_category import ConversationCategory
 from app.models.conversation_scenario import (
     ConversationScenario,
-    ConversationScenarioCreate,
-    ConversationScenarioRead,
     ConversationScenarioStatus,
 )
-from app.models.personalization_option import PersonalizationOptionRead
-from app.models.rating import Rating, RatingCreate, RatingRead
-from app.models.review import Review, ReviewCreate, ReviewRead
+from app.models.review import Review
 from app.models.scenario_preparation import (
     ScenarioPreparation,
-    ScenarioPreparationCreate,
-    ScenarioPreparationRead,
     ScenarioPreparationStatus,
 )
-from app.models.session import (
-    Session,
-    SessionCreate,
-    SessionDetailsRead,
-    SessionRead,
-)
-from app.models.session_feedback import (
-    FeedbackStatusEnum,
-    NegativeExample,
-    PositiveExample,
-    Recommendation,
-    SessionFeedback,
-    SessionFeedbackCreate,
-    SessionFeedbackRead,
-)
-from app.models.session_turn import (
-    SessionTurn,
-    SessionTurnCreate,
-    SessionTurnRead,
-)
+from app.models.session import Session, SessionStatus
+from app.models.session_feedback import FeedbackStatusEnum, SessionFeedback
+from app.models.session_turn import SessionTurn
 from app.models.user_confidence_score import (
-    ConfidenceScoreRead,
     UserConfidenceScore,
-    UserConfidenceScoreCreate,
-    UserConfidenceScoreRead,
 )
-from app.models.user_goal import UserGoal, UserGoalCreate, UserGoalRead
-from app.models.user_profile import (
-    UserProfile,
-    UserProfileExtendedRead,
-    UserProfileRead,
-)
+from app.models.user_goal import UserGoal
+from app.models.user_profile import UserProfile
 
 __all__ = [
     'ConversationCategory',
-    'ConversationCategoryCreate',
-    'ConversationCategoryRead',
     'ConversationScenario',
-    'ConversationScenarioCreate',
-    'ConversationScenarioRead',
     'ConversationScenarioStatus',
     'Session',
-    'SessionCreate',
-    'SessionRead',
-    'SessionDetailsRead',
+    'SessionStatus',
     'ScenarioPreparation',
-    'ScenarioPreparationCreate',
-    'ScenarioPreparationRead',
     'ScenarioPreparationStatus',
     'SessionTurn',
-    'SessionTurnCreate',
-    'SessionTurnRead',
     'SessionFeedback',
-    'SessionFeedbackCreate',
-    'SessionFeedbackRead',
-    'Rating',
-    'RatingCreate',
-    'RatingRead',
     'UserProfile',
-    'UserProfileRead',
     'UserGoal',
-    'UserGoalCreate',
-    'UserGoalRead',
-    'PositiveExample',
     'FeedbackStatusEnum',
-    'NegativeExample',
-    'Recommendation',
     'UserConfidenceScore',
-    'UserConfidenceScoreCreate',
-    'UserConfidenceScoreRead',
-    'ConfidenceScoreRead',
-    'UserProfileExtendedRead',
-    'PersonalizationOptionRead',
     'AppConfig',
-    'AppConfigCreate',
-    'AppConfigRead',
     'ConfigType',
     'AdminDashboardStats',
-    'AdminDashboardStatsRead',
     'Review',
-    'ReviewCreate',
-    'ReviewRead',
 ]
