+++ backend/app/models/conversation_category.py
@@ -1,5 +1,5 @@
 from datetime import UTC, datetime
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING
 
 from sqlalchemy import event
 from sqlalchemy.engine import Connection
@@ -38,30 +38,3 @@ def update_timestamp(
     mapper: Mapper, connection: Connection, target: 'ConversationCategory'
 ) -> None:
     target.updated_at = datetime.now(UTC)
-
-
-# Schema for creating a new ConversationCategory
-class ConversationCategoryCreate(CamelModel):
-    id: str
-    name: Optional[str] = None
-    system_prompt: Optional[str] = None
-    initial_prompt: Optional[str] = None
-    ai_setup: Optional[dict] = Field(default_factory=dict)
-    default_context: Optional[str] = None
-    default_goal: Optional[str] = None
-    default_other_party: Optional[str] = None
-    is_custom: Optional[bool] = None
-    language_code: LanguageCode = Field(default=LanguageCode.en)
-
-
-# Schema for reading ConversationCategory data
-class ConversationCategoryRead(CamelModel):
-    id: str  # Changed from UUID to str
-    name: str
-    default_context: str
-    default_goal: str
-    default_other_party: str
-    is_custom: bool
-    language_code: LanguageCode
-    created_at: datetime
-    updated_at: datetime
