+++ backend/app/models/session.py
@@ -9,11 +9,9 @@ from sqlalchemy.orm.mapper import Mapper
 from sqlmodel import JSON, Column, Field, Relationship
 
 from app.models.camel_case import CamelModel
-from app.models.session_feedback import SessionFeedbackMetrics
 
 if TYPE_CHECKING:
     from app.models.conversation_scenario import ConversationScenario
-    from app.models.rating import Rating
     from app.models.review import Review
     from app.models.session_feedback import SessionFeedback
     from app.models.session_turn import SessionTurn
@@ -42,7 +40,6 @@ class Session(CamelModel, table=True):
     feedback: Optional['SessionFeedback'] = Relationship(
         back_populates='session', cascade_delete=True
     )
-    ratings: list['Rating'] = Relationship(back_populates='session', cascade_delete=True)
     session_review: Optional['Review'] = Relationship(back_populates='session', cascade_delete=True)
     # Automatically update `updated_at` before an update
 
@@ -50,49 +47,3 @@ class Session(CamelModel, table=True):
 @event.listens_for(Session, 'before_update')
 def update_timestamp(mapper: Mapper, connection: Connection, target: 'Session') -> None:
     target.updated_at = datetime.now(UTC)
-
-
-# Schema for creating a new Session
-class SessionCreate(CamelModel):
-    scenario_id: UUID
-    scheduled_at: datetime | None = None
-    started_at: datetime | None = None
-    ended_at: datetime | None = None
-    ai_persona: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    status: SessionStatus = Field(default=SessionStatus.started)
-
-
-class SessionUpdate(CamelModel):
-    scenario_id: UUID | None = None
-    scheduled_at: datetime | None = None
-    started_at: datetime | None = None
-    ended_at: datetime | None = None
-    ai_persona: Optional[dict] = None
-    status: Optional[SessionStatus] = None
-
-
-# Schema for reading Session data
-class SessionRead(CamelModel):
-    id: UUID
-    scenario_id: UUID
-    scheduled_at: datetime | None
-    started_at: datetime | None
-    ended_at: datetime | None
-    ai_persona: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    status: SessionStatus = Field(default=SessionStatus.started)
-    created_at: datetime
-    updated_at: datetime
-
-
-# Schema for reading Session data with details including skill scores, goals achieved,
-# session metrics, and feedback insights
-class SessionDetailsRead(SessionRead):
-    title: str | None = None
-    summary: str | None = None
-    goals_total: list[str] | None = None
-    feedback: Optional['SessionFeedbackMetrics'] = None
-    # List of audio file URIs --> located in session_turns
-    audio_uris: list[str] = Field(default_factory=list)
-
-
-SessionDetailsRead.model_rebuild()
