+++ backend/app/models/session_feedback.py
@@ -9,11 +9,6 @@ from sqlalchemy.orm.mapper import Mapper
 from sqlmodel import JSON, Column, Field, Relationship
 
 from app.models.camel_case import CamelModel
-from app.schemas.session_feedback_schema import (
-    NegativeExample,
-    PositiveExample,
-    Recommendation,
-)
 
 if TYPE_CHECKING:
     from app.models.session import Session
@@ -52,55 +47,3 @@ class SessionFeedback(CamelModel, table=True):  # `table=True` makes it a databa
 @event.listens_for(SessionFeedback, 'before_update')
 def update_timestamp(mapper: Mapper, connection: Connection, target: 'SessionFeedback') -> None:
     target.updated_at = datetime.now(UTC)
-
-
-# Schema for creating a new SessionFeedback
-class SessionFeedbackCreate(CamelModel):
-    session_id: UUID
-    scores: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    tone_analysis: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    overall_score: int
-    transcript_uri: str
-    speak_time_percent: float
-    questions_asked: int
-    session_length_s: int
-    goals_achieved: list[str] = Field(default_factory=list)
-    example_positive: list[PositiveExample] = Field(default_factory=list)
-    example_negative: list[NegativeExample] = Field(default_factory=list)
-    recommendations: list[Recommendation] = Field(default_factory=list)
-    status: FeedbackStatusEnum = FeedbackStatusEnum.pending
-
-
-# Schema for reading SessionFeedback data
-class SessionFeedbackRead(CamelModel):
-    id: UUID
-    session_id: UUID
-    scores: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    tone_analysis: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    overall_score: int
-    transcript_uri: str
-    speak_time_percent: float
-    questions_asked: int
-    session_length_s: int
-    goals_achieved: list[str] = Field(default_factory=list)
-    example_positive: list[PositiveExample] = Field(default_factory=list)
-    example_negative: list[NegativeExample] = Field(default_factory=list)
-    recommendations: list[Recommendation] = Field(default_factory=list)
-    status: FeedbackStatusEnum
-    created_at: datetime
-    updated_at: datetime
-
-
-# Schema for reading a session's feedback metrics
-class SessionFeedbackMetrics(CamelModel):
-    scores: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    tone_analysis: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    overall_score: int
-    transcript_uri: str
-    speak_time_percent: float
-    questions_asked: int
-    session_length_s: int
-    goals_achieved: list[str] = Field(default_factory=list)
-    example_positive: list[PositiveExample] = Field(default_factory=list)
-    example_negative: list[NegativeExample] = Field(default_factory=list)
-    recommendations: list[Recommendation] = Field(default_factory=list)
