+++ backend/app/routers/user_profile_route.py
@@ -7,10 +7,11 @@ from sqlmodel import select
 
 from app.database import get_db_session
 from app.dependencies import require_user
-from app.models.user_confidence_score import ConfidenceScoreRead, UserConfidenceScore
+from app.models.user_confidence_score import UserConfidenceScore
 from app.models.user_goal import Goal, UserGoal
-from app.models.user_profile import (
-    UserProfile,
+from app.models.user_profile import UserProfile
+from app.schemas.user_confidence_score import ConfidenceScoreRead
+from app.schemas.user_profile import (
     UserProfileExtendedRead,
     UserProfileRead,
     UserProfileReplace,
