+++ backend/app/schemas/__init__.py
@@ -1,3 +1,100 @@
-from .message_schema import MessageCreateSchema, MessageSchema
+from app.schemas.admin_dashboard_stats import AdminDashboardStatsRead
+from app.schemas.app_config import AppConfigCreate, AppConfigRead
+from app.schemas.conversation_category import ConversationCategoryRead
+from app.schemas.conversation_scenario import ConversationScenarioCreate, ConversationScenarioRead
+from app.schemas.message_schema import MessageCreateSchema, MessageSchema
+from app.schemas.review import (
+    PaginatedReviewsResponse,
+    ReviewCreate,
+    ReviewRead,
+    ReviewResponse,
+    ReviewStatistics,
+)
+from app.schemas.scenario_preparation import (
+    ChecklistRequest,
+    ConversationScenarioBase,
+    KeyConcept,
+    KeyConceptRequest,
+    KeyConceptResponse,
+    ObjectiveRequest,
+    ScenarioPreparationCreate,
+    ScenarioPreparationRead,
+    StringListResponse,
+)
+from app.schemas.session import (
+    SessionCreate,
+    SessionDetailsRead,
+    SessionRead,
+    SessionUpdate,
+)
+from app.schemas.session_feedback import (
+    ExamplesRequest,
+    GoalsAchievedCollection,
+    GoalsAchievementRequest,
+    NegativeExample,
+    PositiveExample,
+    Recommendation,
+    RecommendationsCollection,
+    RecommendationsRequest,
+    SessionExamplesCollection,
+    SessionFeedbackMetrics,
+)
+from app.schemas.session_turn import (
+    SessionTurnCreate,
+    SessionTurnRead,
+)
+from app.schemas.user_confidence_score import ConfidenceScoreRead
+from app.schemas.user_profile import (
+    UserProfileExtendedRead,
+    UserProfileRead,
+    UserProfileReplace,
+    UserProfileUpdate,
+    UserStatisticsRead,
+)
 
-__all__ = ['MessageSchema', 'MessageCreateSchema']
+__all__ = [
+    'MessageSchema',
+    'MessageCreateSchema',
+    'AdminDashboardStatsRead',
+    'AppConfigCreate',
+    'AppConfigRead',
+    'ConversationCategoryRead',
+    'ConversationScenarioCreate',
+    'ConversationScenarioRead',
+    'ReviewCreate',
+    'ReviewRead',
+    'ReviewResponse',
+    'ReviewStatistics',
+    'PaginatedReviewsResponse',
+    'ChecklistRequest',
+    'ConversationScenarioBase',
+    'KeyConcept',
+    'KeyConceptRequest',
+    'KeyConceptResponse',
+    'ObjectiveRequest',
+    'ScenarioPreparationCreate',
+    'ScenarioPreparationRead',
+    'StringListResponse',
+    'ExamplesRequest',
+    'GoalsAchievedCollection',
+    'GoalsAchievementRequest',
+    'NegativeExample',
+    'PositiveExample',
+    'Recommendation',
+    'RecommendationsCollection',
+    'RecommendationsRequest',
+    'SessionExamplesCollection',
+    'SessionFeedbackMetrics',
+    'SessionTurnCreate',
+    'SessionTurnRead',
+    'SessionCreate',
+    'SessionDetailsRead',
+    'SessionRead',
+    'SessionUpdate',
+    'ConfidenceScoreRead',
+    'UserProfileUpdate',
+    'UserProfileReplace',
+    'UserProfileRead',
+    'UserProfileExtendedRead',
+    'UserStatisticsRead',
+]
